refactor: FSD(Feature-Sliced Design) 아키텍처로 프로젝트 구조 전면 개편
## 주요 변경사항

### 🏗️ FSD 레이어 구조 도입
- **app**: Next.js App Router 및 전역 설정
- **views**: 페이지 컴포넌트 (기존 pages 대체)
- **widgets**: 복합 UI 위젯 (대시보드 컴포넌트)
- **features**: 비즈니스 피처 및 훅
- **entities**: 도메인 로직 (crypto 비즈니스 로직)
- **shared**: 공통 리소스 (api, ui, types, lib)

### 📁 파일 재배치 및 모듈화
- API 레이어: `lib/` → `shared/api/` + `entities/crypto/`
- UI 컴포넌트: `components/` → `shared/ui/` + `widgets/`
- 타입 정의: `types/` → `shared/types/`
- 비즈니스 훅: `hooks/` → `features/dashboard/` + `shared/lib/`

### 🔧 개발자 경험 개선
- Barrel exports를 통한 깔끔한 import 경로
- FSD 원칙에 따른 의존성 방향 통제
- 모듈 간 결합도 감소 및 응집도 증가

### 🚀 빌드 최적화
- 동적 로딩을 통한 SSR 이슈 해결
- TypeScript 엄격 타입 검증 유지
- ESLint 규칙 준수

이 리팩토링으로 코드 유지보수성, 확장성, 가독성이 크게 향상되었습니다.
