+++ src/features/dashboard/useDashboardData.ts
@@ -1,0 +1,1 @@
+export { useDashboardData } from '@/shared/lib/useQueries';
