--- src/hooks/useDashboardData.ts
@@ -1,3 +1,0 @@
-// TanStack Query 기반 useDashboardData는 useQueries.ts로 이동되었습니다.
-// 기존 호환성을 위해 re-export합니다.
-export { useDashboardData } from './useQueries';
