+++ src/shared/types/crypto.ts
