+++ src/shared/ui/ErrorBoundary.tsx
