+++ src/shared/ui/QueryErrorBoundary.tsx
