+++ src/shared/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as ErrorBoundary } from './ErrorBoundary';
+export { default as LoadingSpinner, TableLoadingSpinner, ChartLoadingSpinner } from './LoadingSpinner';
+export { default as QueryErrorBoundary } from './QueryErrorBoundary';
