+++ src/widgets/crypto-dashboard/index.ts
@@ -1,0 +1,2 @@
+export { default as IndicatorChart } from './IndicatorChart';
+export { default as IndicatorTable } from './IndicatorTable';
